Added a long outstanding feature to the cache helper - --cropdetect
The idea is to update each file handler in the fxd files with the mplayer
option for cropping the video Replaced kaa.metadata with mplayer identify in
the encodingcore, metadata doesn't work for all the media types that mplayer
and mencoder can handle. Added the workaround to the encodingcore for mplayer's
broken sstep option.


git-svn-id: svn://svn.freevo.org/freevo/branches/rel-1/freevo@11300 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5