Interim commit fixing the redisplay of the busy icon.
The main problem is that there are two surfaces, the screen and the main_layer.
The screen is what is being displayed and the main_layer is what will be
displayed next. After the icon is stopped there was no restore of the
main_layer which means that part of the icon was still being displayed.

There are still two bugs in the code, positioning and scaling of the backup.
There are a lot of prints that need to be turned into _degug_s


git-svn-id: svn://svn.freevo.org/freevo/branches/rel-1/freevo@11446 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5